requirements.txt
asyncpg==0.28.0
discord.py==2.3.2
parsedatetime==2.6
prettytable==3.9.0
pygit2==1.13.1
pymongo==4.5.0
python_dateutil==2.8.2
pytz==2023.3.post1
RoboArt==0.1.4
setuptools==68.1.2
tabulate==0.9.0
typing_extensions==4.8.0
